Separate ClassLoader and ClassPathAppender
